refactor: EventForm 컴포넌트 분리 & 상수 분리
